+++ src/Accounts/Authentication/Config/Internal/Providers/EnvironmentVariablesConfigurationProvider.cs
@@ -1,0 +1,49 @@
+﻿// ----------------------------------------------------------------------------------
+//
+// Copyright Microsoft Corporation
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// ----------------------------------------------------------------------------------
+
+using Microsoft.Azure.Commands.Common.Authentication.Config.Internal.Interfaces;
+using System;
+using System.Collections.Generic;
+
+namespace Microsoft.Azure.Commands.Common.Authentication.Config.Internal.Providers
+{
+    internal class EnvironmentVariablesConfigurationProvider : ConfigurationProvider
+    {
+        private EnvironmentVariableTarget _environmentVariableTarget;
+        private IDictionary<string, string> _environmentVariableNameToKeyMapping;
+        private IEnvironmentVariableProvider _environmentVariableProvider;
+
+        public EnvironmentVariablesConfigurationProvider(string id, EnvironmentVariablesConfigurationOptions options) : base(id)
+        {
+            _environmentVariableTarget = options.EnvironmentVariableTarget;
+            _environmentVariableNameToKeyMapping = options.EnvironmentVariableToKeyMap ?? new Dictionary<string, string>();
+            _environmentVariableProvider = options.EnvironmentVariableProvider;
+        }
+
+        public override void Load()
+        {
+            var data = new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase);
+            foreach (var i in _environmentVariableNameToKeyMapping)
+            {
+                string value = _environmentVariableProvider.Get(i.Key, _environmentVariableTarget);
+                if (!string.IsNullOrEmpty(value))
+                {
+                    data[i.Value] = value;
+                }
+            }
+
+            Data = data;
+        }
+    }
+}