+++ src/Security/Security/Cmdlets/AlertsSuppressionRules/RemoveAlertsSuppressionRule.cs
@@ -1,0 +1,58 @@
+﻿using System.Management.Automation;
+using Commands.Security;
+using Microsoft.Azure.Commands.Security.Common;
+using Microsoft.Azure.Commands.Security.Models.AlertsSuppressionRules;
+using Microsoft.Azure.Commands.SecurityCenter.Common;
+using Microsoft.Azure.Management.Security;
+
+namespace Microsoft.Azure.Commands.Security.Cmdlets.AlertsSuppressionRules
+{
+    [Cmdlet(VerbsCommon.Remove, ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "AlertsSuppressionRule", DefaultParameterSetName = ParameterSetNames.SubscriptionLevelResource, SupportsShouldProcess = true), OutputType(typeof(bool))]
+    public class RemoveAlertsSuppressionRule : SecurityCenterCmdletBase
+    {
+        [Parameter(ParameterSetName = ParameterSetNames.SubscriptionLevelResource, Mandatory = true, HelpMessage = ParameterHelpMessages.ResourceName)]
+        [ValidateNotNullOrEmpty]
+        public string Name { get; set; }
+
+        [Parameter(ParameterSetName = ParameterSetNames.ResourceId, Mandatory = true, ValueFromPipelineByPropertyName = true, HelpMessage = ParameterHelpMessages.ResourceId)]
+        [ValidateNotNullOrEmpty]
+        public string ResourceId { get; set; }
+
+        [Parameter(ParameterSetName = ParameterSetNames.InputObject, Mandatory = true, ValueFromPipeline = true, HelpMessage = ParameterHelpMessages.InputObject)]
+        [ValidateNotNullOrEmpty]
+        public PSAlertsSuppressionRule InputObject { get; set; }
+
+        [Parameter(Mandatory = false, HelpMessage = ParameterHelpMessages.PassThru)]
+        public SwitchParameter PassThru { get; set; }
+
+        public override void ExecuteCmdlet()
+        {
+            var name = Name;
+
+            switch (ParameterSetName)
+            {
+                case ParameterSetNames.SubscriptionLevelResource:
+                    // name was already set before the switch
+                    break;
+                case ParameterSetNames.ResourceId:
+                    name = AzureIdUtilities.GetResourceName(ResourceId);
+                    break;
+                case ParameterSetNames.InputObject:
+                    name = InputObject.Name;
+                    break;
+                default:
+                    throw new PSInvalidOperationException();
+            }
+
+            if (ShouldProcess(name, VerbsCommon.Remove))
+            {
+                SecurityCenterClient.AlertsSuppressionRules.DeleteWithHttpMessagesAsync(name);
+            }
+
+            if (PassThru.IsPresent)
+            {
+                WriteObject(true);
+            }
+        }
+    }
+}