+++ src/Storage/Storage.Management/ChangeLog.md
@@ -20,6 +20,8 @@
 ## Upcoming Release
 * Show OAuth token in debug log in debug build only
     -  `New-AzStorageContext`
+* Supported return more file properties when list Azure file
+    -  `Get-AzStorageFile`
 
 ## Version 4.5.0
 * Supported DaysAfterLastTierChangeGreaterThan in Management Policy